removed warnings

+++ src/test/scala/com/quadstingray/speedtest/ndt7/SpeedTestSpec.scala
@@ -13,7 +13,7 @@ class SpeedTestSpec extends Specification {
       speedTestResult.download.megaBitPerSecond must beGreaterThan(50.0)
       speedTestResult.upload.megaBitPerSecond must beGreaterThan(5.0)
       speedTestResult.testResults.size must beEqualTo(2)
-      speedTestResult.latency.toLong must beGreaterThanOrEqualTo(1000L)
+      speedTestResult.latency must beGreaterThanOrEqualTo(1000L)
     }
 
     "run Test with given server" >> {
@@ -23,7 +23,7 @@ class SpeedTestSpec extends Specification {
       speedTestResult.download.megaBitPerSecond must beGreaterThan(50.0)
       speedTestResult.upload.megaBitPerSecond must beGreaterThan(5.0)
       speedTestResult.testResults.size must beEqualTo(2)
-      speedTestResult.latency.toLong must beGreaterThanOrEqualTo(1000L)
+      speedTestResult.latency must beGreaterThanOrEqualTo(1000L)
     }
 
     "run Test with not existing server" >> {
@@ -31,7 +31,7 @@ class SpeedTestSpec extends Specification {
       speedTestResult.download.megaBitPerSecond must beGreaterThan(50.0)
       speedTestResult.upload.megaBitPerSecond must beGreaterThan(5.0)
       speedTestResult.testResults.size must beEqualTo(2)
-      speedTestResult.latency.toLong must beGreaterThanOrEqualTo(1000L)
+      speedTestResult.latency must beGreaterThanOrEqualTo(1000L)
     }
 
     "run Test with callbacks" >> {
@@ -54,7 +54,7 @@ class SpeedTestSpec extends Specification {
       speedTestResult.download.megaBitPerSecond must beGreaterThan(50.0)
       speedTestResult.upload.megaBitPerSecond must beGreaterThan(5.0)
       speedTestResult.testResults.size must beEqualTo(2)
-      speedTestResult.latency must beGreaterThan(1000l)
+      speedTestResult.latency must beGreaterThan(1000L)
 
     }
